requirements.txt
blinker==1.9.0
click==8.3.0
cryptography==42.0.8
Flask==3.1.2
flask-socketio==5.5.1
eventlet==0.33.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.0.0
Werkzeug==3.1.3
pytest==8.3.3  
flask-testing==0.8.1
PyMySQL==1.1.0
redis==5.0.1